Sule

+++ src/minidb/basic/database/DataBase.java
@@ -13,11 +13,17 @@ import minidb.result.BoolResult;
 import minidb.result.QueryResult;
 import minidb.result.Result; 
 
+//TODO rhs value of where
 public class DataBase{
 	
 
 	String name;
 	HashMap<String,Table> tables;
+	public DataBase(String n) throws ClassNotFoundException, IOException {
+		tables=new HashMap<String,Table>();
+		name=n;
+		open();
+	}
 	
 	
 	public DataBase() throws ClassNotFoundException, IOException {
@@ -61,15 +67,26 @@ public class DataBase{
 				sa.types=sc.types;
 				this.createTable(sc.tableName,sa);
 			}
+//			else {
+//				throw new IllegalArgumentException("table already exist");
+//			}
 			res=new BoolResult();
 			break;
 		case Statement.drop:
 			StatementDrop sd=(StatementDrop) st;
+			if(!this.tables.containsKey(sd.tableName)) {
+				throw new IllegalArgumentException("table not exist");
+			}
+
 			dropTable(sd.tableName);
 			res=new BoolResult();
 			break;
 		case Statement.insertA:
 			StatementInsertA sia=(StatementInsertA) st;
+			if(!this.tables.containsKey(sia.tableName)) {
+				throw new IllegalArgumentException("table not exist");
+			}
+
 			tb=tables.get(sia.tableName);
 			pair=tb.mkRow(sia.values);
 			tb.simpleInsert(pair.l,pair.r);
@@ -78,6 +95,9 @@ public class DataBase{
 			break;
 		case Statement.insertB:
 			StatementInsertB sib=(StatementInsertB) st;
+			if(!this.tables.containsKey(sib.tableName)) {
+				throw new IllegalArgumentException("table not exist");
+			}
 			tb=tables.get(sib.tableName);
 			pair=tb.mkRowB(sib.pairs);
 			tb.simpleInsert(pair.l,pair.r);
@@ -86,19 +106,34 @@ public class DataBase{
 			break;
 		case Statement.selectA:
 			StatementSelectA sla=(StatementSelectA) st;
+			if(!this.tables.containsKey(sla.tableName)) {
+				throw new IllegalArgumentException("table not exist");
+			}
 			tb=tables.get(sla.tableName);
 			if(sla.isStar) {
 				List<String> names=new ArrayList<String>(tb.schema.descriptors.keySet());
-				res=tb.query(names, sla.existWhere, sla.cdName, sla.cdValue, sla.op);
+				if(sla.isImme)
+					res=tb.query(names, sla.existWhere, sla.cdName, sla.cdValue, sla.op);
+				else
+					res=tb.queryI(names, sla.existWhere, sla.cdName, sla.cdNamer, sla.op);
+					
 			}
 			else
-				res=tb.query(sla.names, sla.existWhere, sla.cdName, sla.cdValue, sla.op);
+				if(sla.isImme)
+					res=tb.query(sla.names, sla.existWhere, sla.cdName, sla.cdValue, sla.op);
+				else
+					res=tb.queryI(sla.names, sla.existWhere, sla.cdName, sla.cdNamer, sla.op);
 			break;
 		case Statement.selectB:
+			StatementSelectB slb=(StatementSelectB) st;
+			res=Table.queryJ(slb.isStar,tables,slb.cnames,slb.jnames,slb.onConditions,slb.existWhere,slb.isImme,slb.cdNameP,slb.cdValue,slb.cdNamerP,slb.op);
 			break;
 
 		case Statement.update:
 			StatementUpdate su=(StatementUpdate) st;
+			if(!this.tables.containsKey(su.tableName)) {
+				throw new IllegalArgumentException("table not exist");
+			}
 			tb=tables.get(su.tableName);
 			tb.update(su.cdName,su.cdValue,su.op,su.setName,su.setValue);
 			res=new BoolResult();
@@ -106,6 +141,9 @@ public class DataBase{
 
 		case Statement.delete:
 			StatementDelete sds=(StatementDelete) st;
+			if(!this.tables.containsKey(sds.tableName)) {
+				throw new IllegalArgumentException("table not exist");
+			}
 			tb=tables.get(sds.tableName);
 			tb.delete(sds.cdName,sds.cdValue,sds.op);
 			res=new BoolResult();
@@ -125,9 +163,16 @@ public class DataBase{
 	}
 	
 	public void dropTable(String name) throws IOException {
+		Table tb=tables.get(name);
 		tables.remove(name);
+		for(String n:tb.schema.descriptors.keySet()) {
+			File filef = new File(name.concat("_").concat(n).concat(".index"));
+			filef.delete();
+		}
 		File file = new File(name.concat(".schema"));
 		file.delete();
+		File file1 = new File(name.concat(".index"));
+		file1.delete();
 		File file2 = new File("schema.log");
 		file2.delete();
 		refreshLog();
